Use unsigned long to read random seed

+++ src/jams/core/jams++.cc
@@ -104,7 +104,7 @@ void jams_initialize(int argc, char **argv) {
   try {
     if (::config->exists("sim")) {
       simulation.verbose = jams::config_optional<bool>(config->lookup("sim"), "verbose", false);
-      simulation.random_seed = jams::config_optional<unsigned>(config->lookup("sim"), "seed", simulation.random_seed);
+      simulation.random_seed = jams::config_optional<unsigned long>(config->lookup("sim"), "seed", simulation.random_seed);
       if (config->exists("sim.rng_state")) {
         auto state = jams::config_required<string>(config->lookup("sim"), "rng_state");
         istringstream(state) >> jams::random_generator();
